move Address toIpSpace into Conversions class
