+++ projects/batfish/src/main/java/org/batfish/representation/fortios/Address.java
@@ -1,18 +1,11 @@
 package org.batfish.representation.fortios;
 
 import static com.google.common.base.MoreObjects.firstNonNull;
-import static com.google.common.base.Preconditions.checkState;
 
 import java.io.Serializable;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import org.batfish.common.Warnings;
-import org.batfish.datamodel.EmptyIpSpace;
 import org.batfish.datamodel.Ip;
-import org.batfish.datamodel.IpRange;
-import org.batfish.datamodel.IpSpace;
-import org.batfish.datamodel.IpWildcard;
-import org.batfish.datamodel.Prefix;
 
 /** FortiOS datamodel component containing address configuration */
 public class Address implements FortiosRenameableObject, Serializable {
@@ -109,52 +102,6 @@ public class Address implements FortiosRenameableObject, Serializable {
     _name = name;
     _typeSpecificFields = new TypeSpecificFields();
     _uuid = uuid;
-  }
-
-  public IpSpace toIpSpace(Warnings w) {
-    // TODO Investigate & support _allowRouting, _associatedInterface, _fabricObject
-    switch (getTypeEffective()) {
-      case IPMASK:
-        Ip subnetIp = getTypeSpecificFields().getIp1Effective();
-        Ip subnetMask = getTypeSpecificFields().getIp2Effective();
-        // Throw if mask is invalid; such an address should not have made it through extraction
-        checkState(
-            subnetMask.isValidNetmask1sLeading(),
-            String.format("Cannot convert address %s: %s is an invalid mask", _name, subnetMask));
-        return Prefix.create(subnetIp, subnetMask).toIpSpace();
-      case IPRANGE:
-        Ip startIp = getTypeSpecificFields().getIp1Effective();
-        Ip endIp = getTypeSpecificFields().getIp2Effective();
-        // Throw if end IP is zero; such an address should not have made it through extraction.
-        // ("end IP cannot be 0" is the warning the CLI gives when end-ip was not set.)
-        checkState(
-            !endIp.equals(Ip.ZERO),
-            String.format("Cannot convert address %s: end IP cannot be 0", _name));
-        // Shouldn't have made it through extraction if end IP > start IP; let range throw if so
-        return IpRange.range(startIp, endIp);
-      case WILDCARD:
-        Ip ip = getTypeSpecificFields().getIp1Effective();
-        // Invert mask because IpWildcard interprets set bits as "don't care", whereas FortiOS
-        // interprets unset bits as "don't care"
-        Ip mask = getTypeSpecificFields().getIp2Effective().inverted();
-        return IpWildcard.ipWithWildcardMask(ip, mask).toIpSpace();
-      case INTERFACE_SUBNET:
-        // TODO test what IPs this actually includes. Docs say it will:
-        //  "automatically create an address object that matches the interface subnet"
-        //  but it's unclear because it supports both "set subnet" and "set interface".
-      case DYNAMIC: // Based on SDN connectors, whose addresses aren't known statically
-      case FQDN: // Based on domain names
-      case GEOGRAPHY: // Based on countries
-      case MAC: // Based on MAC addresses
-        // Unsupported address types.
-        w.redFlag(
-            String.format(
-                "Addresses of type %s are unsupported and will be considered unmatchable.",
-                getType()));
-        return EmptyIpSpace.INSTANCE;
-      default:
-        throw new IllegalStateException("Unrecognized address type " + getTypeEffective());
-    }
   }
 
   public @Nullable Boolean getAllowRouting() {
