+++ projects/batfish/src/test/java/org/batfish/representation/fortios/FortiosConfigurationTest.java
@@ -1,27 +1,42 @@
 package org.batfish.representation.fortios;
 
+import static org.batfish.common.matchers.WarningMatchers.hasText;
+import static org.batfish.representation.fortios.FortiosPolicyConversions.toIpSpace;
 import static org.batfish.representation.fortios.FortiosPolicyConversions.toMatchExpr;
 import static org.batfish.representation.fortios.FortiosTraceElementCreators.matchServiceGroupTraceElement;
 import static org.batfish.representation.fortios.FortiosTraceElementCreators.matchServiceTraceElement;
 import static org.batfish.representation.fortios.Service.DEFAULT_SOURCE_PORT_RANGE;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.Matchers.contains;
+import static org.hamcrest.Matchers.empty;
 import static org.hamcrest.Matchers.equalTo;
 
 import com.google.common.base.Functions;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.ImmutableSet;
 import java.util.Set;
+import java.util.stream.Stream;
 import net.sf.javabdd.BDD;
 import org.batfish.common.Warnings;
 import org.batfish.common.bdd.HeaderSpaceToBDD;
 import org.batfish.common.bdd.IpAccessListToBdd;
 import org.batfish.datamodel.BddTestbed;
+import org.batfish.datamodel.EmptyIpSpace;
 import org.batfish.datamodel.HeaderSpace;
 import org.batfish.datamodel.IntegerSpace;
+import org.batfish.datamodel.Ip;
 import org.batfish.datamodel.IpProtocol;
+import org.batfish.datamodel.IpRange;
+import org.batfish.datamodel.IpSpace;
+import org.batfish.datamodel.IpWildcard;
+import org.batfish.datamodel.Prefix;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.ExpectedException;
 
 public class FortiosConfigurationTest {
+
+  @Rule public ExpectedException _thrown = ExpectedException.none();
 
   private final BddTestbed _bddTestbed = new BddTestbed(ImmutableMap.of(), ImmutableMap.of());
   private final BDD _one = _bddTestbed.getPkt().getFactory().one();
@@ -282,5 +297,117 @@ public class FortiosConfigurationTest {
                             ServiceGroupMember::getName, Functions.identity())),
                 null)),
         equalTo(expected));
+  }
+
+  private static Address createAddressWithIps(Ip ip1, Ip ip2) {
+    Address address = new Address("name", new BatfishUUID(1));
+    address.getTypeSpecificFields().setIp1(ip1);
+    address.getTypeSpecificFields().setIp2(ip2);
+    return address;
+  }
+
+  @Test
+  public void testToIpSpace_ipmaskDefault() {
+    Address address = new Address("name", new BatfishUUID(1));
+    assertConvertsWithoutWarnings(address, Prefix.ZERO.toIpSpace());
+  }
+
+  @Test
+  public void testToIpSpace_ipmask() {
+    Ip ip = Ip.parse("1.1.1.1");
+    Ip mask = Ip.parse("255.255.255.0");
+    Address address = createAddressWithIps(ip, mask);
+    assertConvertsWithoutWarnings(address, Prefix.create(ip, mask).toIpSpace());
+  }
+
+  @Test
+  public void testToIpSpace_ipmaskInvalidMask() {
+    // 1.1.1.0 isn't a valid subnet mask
+    Address address = createAddressWithIps(Ip.parse("1.1.1.0"), Ip.parse("1.1.1.0"));
+    _thrown.expect(IllegalStateException.class);
+    toIpSpace(address, new Warnings());
+  }
+
+  @Test
+  public void testToIpSpace_iprangeDefault() {
+    // Must set end IP
+    Ip endIp = Ip.parse("1.1.1.255");
+    Address address = new Address("name", new BatfishUUID(1));
+    address.setType(Address.Type.IPRANGE);
+    address.getTypeSpecificFields().setIp2(endIp);
+    assertConvertsWithoutWarnings(address, IpRange.range(Ip.ZERO, endIp));
+  }
+
+  @Test
+  public void testToIpSpace_iprange() {
+    Ip startIp = Ip.parse("1.1.1.0");
+    Ip endIp = Ip.parse("1.1.1.255");
+    Address address = createAddressWithIps(startIp, endIp);
+    address.setType(Address.Type.IPRANGE);
+    assertConvertsWithoutWarnings(address, IpRange.range(startIp, endIp));
+  }
+
+  @Test
+  public void testToIpSpace_iprangeNoEndIp() {
+    // 1.1.1.0 isn't a valid subnet mask
+    Address address = new Address("name", new BatfishUUID(1));
+    address.setType(Address.Type.IPRANGE);
+    _thrown.expect(IllegalStateException.class);
+    toIpSpace(address, new Warnings());
+  }
+
+  @Test
+  public void testToIpSpace_iprangeEndIpTooLow() {
+    Address address = createAddressWithIps(Ip.parse("2.2.2.2"), Ip.parse("1.1.1.1"));
+    address.setType(Address.Type.IPRANGE);
+    _thrown.expect(IllegalArgumentException.class);
+    toIpSpace(address, new Warnings());
+  }
+
+  @Test
+  public void testToIpSpace_wildcardDefault() {
+    Address address = new Address("name", new BatfishUUID(1));
+    address.setType(Address.Type.WILDCARD);
+    assertConvertsWithoutWarnings(address, IpWildcard.ANY.toIpSpace());
+  }
+
+  @Test
+  public void testToIpSpace_wildcard() {
+    Ip mask = Ip.parse("255.0.255.128"); // FortiOS format (bits that are set matter)
+    IpWildcard wildcard = IpWildcard.ipWithWildcardMask(Ip.parse("1.1.1.1"), mask.inverted());
+    Address address = createAddressWithIps(wildcard.getIp(), mask);
+    address.setType(Address.Type.WILDCARD);
+    assertConvertsWithoutWarnings(address, wildcard.toIpSpace());
+  }
+
+  @Test
+  public void testToIpSpace_unsupportedTypes() {
+    Stream.of(
+            Address.Type.INTERFACE_SUBNET,
+            Address.Type.DYNAMIC,
+            Address.Type.FQDN,
+            Address.Type.GEOGRAPHY,
+            Address.Type.MAC)
+        .forEach(
+            unsupportedType -> {
+              Address address = new Address("name", new BatfishUUID(1));
+              address.setType(unsupportedType);
+              Warnings w = new Warnings(true, true, true);
+              assertThat(toIpSpace(address, w), equalTo(EmptyIpSpace.INSTANCE));
+              assertThat(
+                  w.getRedFlagWarnings(),
+                  contains(
+                      hasText(
+                          String.format(
+                              "Addresses of type %s are unsupported and will be considered"
+                                  + " unmatchable.",
+                              unsupportedType))));
+            });
+  }
+
+  private static void assertConvertsWithoutWarnings(Address address, IpSpace expected) {
+    Warnings w = new Warnings(true, true, true);
+    assertThat(toIpSpace(address, w), equalTo(expected));
+    assertThat(w.getRedFlagWarnings(), empty());
   }
 }
